Split out MatrixCU and VectorCU
